+++ src/app/domain/models/resource_identity/realms/__init__.py
@@ -1,0 +1,18 @@
+from .local import LocalAddress, LocalCoordinate
+from .network import NetworkAddress, NetworkCoordinate
+from .memory import MemoryAddress, MemoryCoordinate
+from .synthetic import SyntheticAddress, SyntheticCoordinate
+from .virtual import VirtualAddress, VirtualCoordinate
+
+__all__ = [
+    "LocalAddress",
+    "LocalCoordinate",
+    "NetworkAddress",
+    "NetworkCoordinate",
+    "MemoryAddress",
+    "MemoryCoordinate",
+    "SyntheticAddress",
+    "SyntheticCoordinate",
+    "VirtualAddress",
+    "VirtualCoordinate",
+]